requirements.txt
flask==3.0.0
flask-wtf==1.2.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-talisman==1.1.0
flask-limiter==3.5.0
flask-cors==4.0.0
requests==2.31.0
python-bitcoinrpc==1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
qrcode==7.4.2
pillow==10.1.0
